+++ src/routes.ts
@@ -14,6 +14,8 @@ import { CreateOrderController } from './controllers/order/CreateOrderController
 import { RemoveOrderController } from './controllers/order/RemoveOrderController';
 import { AdditemController } from './controllers/order/AdditemController';
 import { RemoveItemController } from './controllers/order/RemoveItemController';
+import { SendOrderController } from './controllers/order/SendOrderController';
+import { DetailOrderController } from './controllers/order/DetailOrderController';
 
 const router = Router();
 
@@ -40,6 +42,7 @@ router.post('/order', isAuthenticated, new CreateOrderController().handle)
 router.delete('/order', isAuthenticated, new RemoveOrderController().handle)
 router.post('/order/add', isAuthenticated, new AdditemController().handle)
 router.delete('/order/remove', isAuthenticated, new RemoveItemController().handle)
-
+router.put('/order/send', isAuthenticated, new SendOrderController().handle)
+router.get('/order', isAuthenticated, new DetailOrderController().handle)
 
 export { router };