+++ src/services/order/DetailOrderService.ts
@@ -1,0 +1,21 @@
+import prismaClient from "../../prisma";
+interface DetailRequest{
+    order_id: string;
+}
+
+class DetailOrderService{
+    async execute({order_id}: DetailRequest){
+        const orders = await prismaClient.item.findFirst({
+            where:{
+                order_id: order_id
+            },
+            include:{
+                veicule:true,
+                order: true
+            }
+        })
+        return orders;
+    }
+}
+
+export { DetailOrderService }